+++ lib/src/models/h3_exception.dart
@@ -1,0 +1,10 @@
+class H3Exception implements Exception {
+  H3Exception(this.message);
+
+  final String message;
+
+  @override
+  String toString() {
+    return "H3Exception: $message";
+  }
+}